backend/requirements.txt
aiosmtplib==5.0.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
blinker==1.9.0
cffi==2.0.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi-mail==1.6.1
greenlet==3.3.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.1.1
regex==2025.11.3
SQLAlchemy==2.0.45
starlette==0.51.0
typing-inspection==0.4.2
typing_extensions==4.15.0
